+++ src/main/java/nl/basmens/raymarchingplayground/engine/Scene.java
@@ -10,5 +10,5 @@ public abstract class Scene {
 
 
   public abstract void init();
-  public abstract void update(float dt);
+  public abstract void update(double dt);
 }
