+++ src/main/java/nl/basmens/raymarchingplayground/util/time.java
@@ -1,9 +1,9 @@
 package nl.basmens.raymarchingplayground.util;
 
 public class Time {
-  public static float timeStarted = System.nanoTime();
+  public static double timeStarted = System.nanoTime();
 
-  public static float getTime() {
-    return (float)((System.nanoTime() - timeStarted) * 1E-9);
+  public static double getTime() {
+    return (System.nanoTime() - timeStarted) * 1E-9;
   }
 }
